Remove the details from the returned payload.
We have been getting a Payload Too Large error when
posting a large amount of text to the quality metrics
service.

This is due to returning the details of each quality error
in the response. We don't make use of this level of detail.

This commit removes the 'messages' array from the response
so the response is now in the following format:

{
  "readability": 2,
  "passive": 1,
  "repeated_words": 3,
  and so on...
}