Increase the bodyParser limit to 50mb
We have been experiencing 'Payload too large' errors from
this service.

The previous attempt to fix this reduced the size of the returned
payload. This was not the problem. The problem was, in reality,
the size of the request payload, which defaults to 100k.

We have now increased this to 50mb.